refactor: split skills executor modules
