+++ agently/builtins/agent_extensions/SkillsExtension/__init__.py
@@ -12,15 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .AgentlySkillsExecutor import AgentlySkillsExecutor
-from .AgentlySkillsExecutor.modules.errors import SkillError, SkillExecutionError, SkillInstallError, SkillNormalizationError
-from .AgentlySkillsExecutor.modules.executor import SkillExecution
+from .SkillsExtension import SkillsExtension
 
-__all__ = [
-    "AgentlySkillsExecutor",
-    "SkillError",
-    "SkillExecution",
-    "SkillExecutionError",
-    "SkillInstallError",
-    "SkillNormalizationError",
-]
+__all__ = ["SkillsExtension"]
