+++ agently/types/plugins/__init__.py
@@ -27,7 +27,7 @@ from .EventHooker import EventHooker
 from .PromptGenerator import PromptGenerator
 from .ModelRequester import ModelRequester
 from .TaskDAGPlanner import TaskDAGPlanner
-from .SkillsExecutor import SkillsExecutorPlugin
+from .SkillsExecutor import SkillsExecutionContext, SkillsExecutor, SkillsPlanningContext, SkillsRuntimeContext
 from .AgentOrchestrator import AgentOrchestrator
 from .ResponseParser import ResponseParser
 from .ToolManager import ToolManager
